resume-analyzer/requirements.txt
PyPDF2
streamlit
google-generativeai
python-dotenv